[Cleanup beta] Move test for Ember.set to observer_test
Following this TODO:
https://github.com/emberjs/ember.js/pull/12491#issuecomment-149023771
